requirements.txt
Django==1.10.7
psycopg2==2.6.2
requests==2.17.3
django-haystack==2.5.0
GDAL==1.10.0
